fix: make renames work correctly
